refactor: domain 성격의 클래스들 패키지 이동
