+++ src/main/kotlin/lotto/view/OutputView.kt
@@ -1,8 +1,8 @@
 package lotto.view
 
-import lotto.Lotto
-import lotto.LottoResult
-import lotto.Prize
+import lotto.domain.Lotto
+import lotto.domain.LottoResult
+import lotto.domain.Prize
 
 class OutputView {
     fun showInputMoneyMessage() {
